# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guest-place",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analytics/engagement": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_engagement_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Engagement stats by entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_EngagementStatsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/funnel": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_funnel_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversion funnel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_FunnelStepDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/referrals": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_referral_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Referral sources",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_ReferralDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/session": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "operationId": "start_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SessionResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/summary": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_summary_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AnalyticsSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/track": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "operationId": "track_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackEventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Event tracked"
          }
        }
      }
    },
    "/api/v1/analytics/traffic": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_traffic_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Traffic stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TrafficStatsDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_AnalyticsSummaryDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "total_visitors",
              "avg_session_duration",
              "total_clicks",
              "page_views",
              "visitors_trend",
              "session_trend",
              "clicks_trend",
              "views_trend"
            ],
            "properties": {
              "avg_session_duration": {
                "type": "integer",
                "format": "int64"
              },
              "clicks_trend": {
                "type": "number",
                "format": "float"
              },
              "page_views": {
                "type": "integer",
                "format": "int64"
              },
              "session_trend": {
                "type": "number",
                "format": "float"
              },
              "total_clicks": {
                "type": "integer",
                "format": "int64"
              },
              "total_visitors": {
                "type": "integer",
                "format": "int64"
              },
              "views_trend": {
                "type": "number",
                "format": "float"
              },
              "visitors_trend": {
                "type": "number",
                "format": "float"
              }
            }
          },
          "errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "messages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ApiResponse_SessionResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          "errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "messages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApiResponse_TrafficStatsDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "labels",
              "dau",
              "sessions"
            ],
            "properties": {
              "dau": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              },
              "labels": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "sessions": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "messages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ApiResponse_Vec_EngagementStatsDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "entity_id",
                "clicks",
                "views"
              ],
              "properties": {
                "clicks": {
                  "type": "integer",
                  "format": "int64"
                },
                "entity_id": {
                  "type": "string"
                },
                "views": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "messages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApiResponse_Vec_FunnelStepDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "name",
                "count",
                "percentage"
              ],
              "properties": {
                "count": {
                  "type": "integer",
                  "format": "int64"
                },
                "name": {
                  "type": "string"
                },
                "percentage": {
                  "type": "number",
                  "format": "float"
                }
              }
            }
          },
          "errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "messages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ApiResponse_Vec_ReferralDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "source",
                "count",
                "percentage"
              ],
              "properties": {
                "count": {
                  "type": "integer",
                  "format": "int64"
                },
                "percentage": {
                  "type": "number",
                  "format": "float"
                },
                "source": {
                  "type": "string"
                }
              }
            }
          },
          "errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "messages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "StartSessionDto": {
        "type": "object",
        "required": [
          "visitor_id"
        ],
        "properties": {
          "referer": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_agent": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "visitor_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "TrackEventDto": {
        "type": "object",
        "required": [
          "session_id",
          "event_type"
        ],
        "properties": {
          "entity_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "entity_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "event_type": {
            "type": "string"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          },
          "properties": {},
          "session_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "gp_apikey"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}